% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_change_side_data.R
\name{convert_and_compare}
\alias{convert_and_compare}
\title{Master run
Do the lot!}
\usage{
convert_and_compare(input_file, output_location = "", save_raw = F)
}
\arguments{
\item{input_file}{a text file}

\item{output_location}{output location}

\item{save_raw}{if T then exports raw as well as summarised}
}
\value{
all the data
}
\description{
Master run
Do the lot!
}
