% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_change_side_data.R
\name{create_change_side_data}
\alias{create_change_side_data}
\title{create change side data}
\usage{
create_change_side_data(
  data,
  window_start = 1750,
  window_stop = 6750,
  side = "side",
  trackloss = 0.75,
  time_column = "Timestamp",
  aoi_columns = c("Target", "Distractor"),
  trackloss_column = "Trackloss",
  participant_column = "ID",
  trial_column = "UTrial"
)
}
\arguments{
\item{window_start}{start time}

\item{window_stop}{stop time}

\item{side}{side column}

\item{trackloss}{trackloss prop, default .75}
}
\value{
a summary output
}
\description{
create change side data
}
