% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_change_side_data.R
\name{create_change_side_data2}
\alias{create_change_side_data2}
\title{create change side data2}
\usage{
create_change_side_data2(
  data,
  window_start = 1750,
  window_stop = 6750,
  side = "side",
  trackloss = 0.75,
  time_column = "timestamp_new",
  aoi_columns = c("target", "distractor"),
  trackloss_column = "trackloss",
  participant_column = "ID",
  trial_column = "trial",
  summary = TRUE
)
}
\arguments{
\item{window_start}{start time}

\item{window_stop}{stop time}

\item{side}{side column}

\item{trackloss}{trackloss prop, default .75}
}
\value{
a summary output
}
\description{
Comes from robocode data only
}
