% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datavyu_trial_data.R
\name{create_datavyu_trial_data}
\alias{create_datavyu_trial_data}
\title{create a trial report from a datavyu output data}
\usage{
create_datavyu_trial_data(data, write = F)
}
\arguments{
\item{data}{a dataframe read in from a datavyu report CSV or txt}

\item{write}{if TRUE will save a csv in the current working directory}
}
\value{
A formatted dataframe with CP, SwitchRate, MLD and TLT, as well as coding info.
}
\description{
This was designed to work with eyelink fixation reports and the VWM trial.
I can't guarantee it will bring out what you want beyond that so please check
the output carefully.
}
\examples{
library(readr)
data <- read_csv("datavyudata.csv")
data_out <- create_datavyu_trial_data(data, write = F)

}
