% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datavyu_trial_data.R
\name{create_datavyu_trial_data2}
\alias{create_datavyu_trial_data2}
\title{make from a cleaned DV file}
\usage{
create_datavyu_trial_data2(data, write = F)
}
\arguments{
\item{write}{if TRUE will save a csv in the current working directory}
}
\value{
A formatted dataframe with CP, SwitchRate, MLD and TLT, as well as coding
}
\description{
#' @param data a dataframe read in from a datavyu report CSV or txt
}
\examples{
library(readr)
data <- read_csv("datavyudata.csv")
data_out <- create_datavyu_trial_data2(data, write = F)

}
