% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downsample.R
\name{downsample}
\alias{downsample}
\title{Downsample a frame-by-frame eyetracking output - ie a sample report}
\usage{
downsample(data, time_bin_size, sample_rate = 2)
}
\arguments{
\item{data}{an eyetracking sample report}

\item{time_bin_size}{the size of the desired timebin, in ms}

\item{sample_rate}{the current sample rate of the original output}
}
\value{
A downsampled eyetracking sample report
}
\description{
This downsamples the values, taking means across time bins
}
\examples{
library(readr)
data <- read_csv('eyetracking_sample.csv')
data_out <- downsample(data, 100, 2)

}
