% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kappa_Test.R
\name{get_kappa_value}
\alias{get_kappa_value}
\title{Kappa from ML}
\usage{
get_kappa_value(data, start_time = 0, stop_time = 10000)
}
\arguments{
\item{data}{a dataset ready for ML and ET}

\item{start}{a start time}

\item{stop}{a stop time}
}
\value{
a kappa value
}
\description{
Kappa from ML
}
