% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_change_side_data.R
\name{join_CS_to_VE}
\alias{join_CS_to_VE}
\title{Join a change side data with a visexp data}
\usage{
join_CS_to_VE(change_data, visexp_data, .other = NULL)
}
\arguments{
\item{change_data}{change side dataframe}

\item{visexp_data}{vis exp dataframe}

\item{.extra}{other columns to keep}
}
\value{
a clean dataframe
}
\description{
Join a change side data with a visexp data
}
