% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_ET_datavyu.R
\name{merge_ET_datavyu}
\alias{merge_ET_datavyu}
\title{Merge eyetracking and datavyu data}
\usage{
merge_ET_datavyu(ETdata, datavyudata, write = FALSE)
}
\arguments{
\item{ETdata}{a file of processed eyetracking data}

\item{datavyudata}{processed output from datavyu file}

\item{write}{if TRUE will write a file directly to the current wd}
}
\value{
a merged dataframe where eyetracking data is replaced with look coded data where necessary
}
\description{
Both files should first have been processed with the relevant commands
}
\examples{
library(readr)
data <- read_csv("etdata.csv")
data_out <- create_ET_trial_data(data, task = 'VWM', write = F)
data2 <- read_csv("datavyudata.csv")
data2_out <- create_datavyu_trial_data(data2, write = F)
data3 <- merge(data_out, data2_out, write = F)

}
