% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_both.R
\name{read_et_data}
\alias{read_et_data}
\title{read eyetracking data from filepath}
\usage{
read_et_data(filepath, sep = "\\t")
}
\arguments{
\item{filepath}{a quoted filepath}

\item{sep}{the delimeter, defaults to tab}
}
\value{
a rectangular dataframe from the eyetracking output
}
\description{
read eyetracking data from filepath
}
\examples{
\dontrun{
et_data <- read_et_data('data/eyetracking.txt', sep = '\t')
}
}
