% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_both.R
\name{read_video_data}
\alias{read_video_data}
\title{read video data from filepath}
\usage{
read_video_data(filepath)
}
\arguments{
\item{filepath}{a quoted filepath}
}
\value{
a dataframe with annotation
}
\description{
read video data from filepath
}
\examples{
 \dontrun{
video_data <- read_video_data('data/annotate.txt')
}
}
