% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kappa_Test.R
\name{reliability_test}
\alias{reliability_test}
\title{Run a kappa test on preferential looking data to test for reliability in video coding from datavyu}
\usage{
reliability_test(data, Subject1, Subject2, Trials)
}
\arguments{
\item{data}{a data frame, or preferentially a tibble}

\item{Subject1}{a column of directions coding by Subject 1}

\item{Subject2}{a column of directions coding by Subject 2}

\item{Trials}{a column of}
}
\value{
Kappa test output of coding reliability
}
\description{
Run a kappa test on preferential looking data to test for reliability in video coding from datavyu
}
\examples{
data <- read_csv(mydata.csv)
reliability_test(data, 'LOOKING.DIRECTION', 'Looking_JG2.DIRECTION', 'TRIALS.SS')

}
