% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallet.R
\name{run_mallet_data}
\alias{run_mallet_data}
\title{run mallet data
runs a mallet processing script on a path. If write is true it will write the output}
\usage{
run_mallet_data(path, write = T)
}
\arguments{
\item{path}{a file path from mallet}

\item{write}{write output files? default is true}
}
\value{
a list of files: summary, vwm, pingu
}
\description{
run mallet data
runs a mallet processing script on a path. If write is true it will write the output
}
