% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_polhemus.R
\name{transform_polhemus}
\alias{transform_polhemus}
\title{transform polhemus data into digpoints for AtlasViewerGUI}
\usage{
transform_polhemus(filename, wd, option = c("NIH", "Gates", "HWB"),
  out_folder = "DigTest")
}
\arguments{
\item{filename}{A string denoting name of the file}

\item{wd}{A string denoting working directory}

\item{option}{Experiment setup cap configuration, from NIH, Gates, or HWB}

\item{out_folder}{A string denoting out folder name, by default within wd}
}
\value{
A saved digpts file in the out_folder
}
\description{
This works with NIH, Gates, and HWB data. If you have different data then contact Sam
}
\examples{
transform_polhemus('06NIHVWM111B', 'Desktop/Digitization/'m option = 'NIH', out_folder = 'DigTest')

}
