% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_change_side_data.R
\name{convert_and_compare}
\alias{convert_and_compare}
\title{Master run
Do the lot!}
\usage{
convert_and_compare(
  input_file,
  output_location = "",
  save_raw = F,
  task = "VWM",
  .kappa_start = 0,
  .kappa_stop = 10000
)
}
\arguments{
\item{input_file}{a text file}

\item{output_location}{output location}

\item{save_raw}{if T then exports raw as well as summarised}

\item{task}{a task}

\item{.kappa_start}{pass to get_kappa}

\item{.kappa_stop}{pass to get_kappa}
}
\value{
all the data
}
\description{
Master run
Do the lot!
}
